refactor: use riskEngine market listed event for pToken creation
