+++ src/factory.ts
@@ -1,15 +1,14 @@
 import { ponder } from "ponder:registry";
 import { protocol, pToken } from "ponder:schema";
 import { getOrCreateTx } from "./utils/transaction";
-import { getProtocolId, getTxId, getUniqueContractId } from "./utils/id";
+import { getTxId, getUniqueContractId } from "./utils/id";
 import { readPTokenInfo } from "./utils/multicalls";
 import { getOrCreateUnderlying } from "./utils/underlying";
 
 ponder.on("Factory:ProtocolDeployed", async ({ context, event }) => {
-  const id = getProtocolId(
-    event.log.address,
+  const id = getUniqueContractId(
     context.network.chainId,
-    event.args.protocolId
+    event.args.riskEngine
   );
   const creationTransactionId = getTxId(event, context);
 
@@ -28,59 +27,13 @@ ponder.on("Factory:ProtocolDeployed", async ({ context, event }) => {
 });
 
 ponder.on("Factory:PTokenDeployed", async ({ context, event }) => {
-  const id = getUniqueContractId(context.network.chainId, event.args.pToken);
-
-  const protocolIdDb = getProtocolId(
-    event.log.address,
-    context.network.chainId,
-    event.args.protocolId
-  );
-
-  const protocolData = await context.db.find(protocol, {
-    id: protocolIdDb,
-  });
-
-  if (!protocolData) {
-    throw new Error("Protocol not found");
-  }
-
-  const riskEngine = protocolData.riskEngine;
-
-  const pTokenInfo = await readPTokenInfo(
-    context,
-    event.args.pToken,
-    riskEngine
-  );
-
-  const underlyingId = getUniqueContractId(
-    context.network.chainId,
-    pTokenInfo.asset
-  );
-
-  await Promise.all([
-    context.db.insert(pToken).values({
-      id,
-      creationTransactionId: getTxId(event, context),
-      chainId: BigInt(context.network.chainId),
-      address: event.args.pToken,
-      protocolId: protocolIdDb,
+  // at this point the pToken entry was already created by the MarketListed event
+  // we just have to include the index data
+  await context.db
+    .update(pToken, {
+      id: getUniqueContractId(context.network.chainId, event.args.pToken),
+    })
+    .set({
       index: event.args.index,
-      underlyingId: underlyingId,
-      symbol: pTokenInfo.symbol,
-      name: pTokenInfo.name,
-      decimals: `${pTokenInfo.decimals}`,
-      borrowCap: pTokenInfo.borrowCap,
-      supplyCap: pTokenInfo.supplyCap,
-      collateralFactor: pTokenInfo.collateralFactor,
-      closeFactor: pTokenInfo.closeFactor,
-      exchangeRateCurrent: pTokenInfo.exchangeRateCurrent,
-      liquidationIncentive: pTokenInfo.liquidationIncentive,
-      liquidationThreshold: pTokenInfo.liquidationThreshold,
-      protocolSeizeShare: pTokenInfo.protocolSeizeShare,
-      borrowRatePerSecond: pTokenInfo.borrowRatePerSecond,
-      supplyRatePerSecond: pTokenInfo.supplyRatePerSecond,
-    }),
-    getOrCreateUnderlying(pTokenInfo.asset, context),
-    getOrCreateTx(event, context),
-  ]);
+    });
 });
